+++ BambooTracker/command/pattern/set_volume_to_step_command.cpp
@@ -1,14 +1,15 @@
 #include "set_volume_to_step_command.hpp"
 #include "misc.hpp"
 
-SetVolumeToStepCommand::SetVolumeToStepCommand(std::weak_ptr<Module> mod, int songNum, int trackNum, int orderNum, int stepNum, int volume)
+SetVolumeToStepCommand::SetVolumeToStepCommand(std::weak_ptr<Module> mod, int songNum, int trackNum, int orderNum, int stepNum, int volume, bool isFMReversed)
 	: mod_(mod),
 	  song_(songNum),
 	  track_(trackNum),
 	  order_(orderNum),
 	  step_(stepNum),
 	  vol_(volume),
-	  isComplete_(false)
+	  isComplete_(false),
+	  isFMReserved_(isFMReversed)
 {
 	prevVol_ = mod_.lock()->getSong(songNum).getTrack(trackNum)
 			   .getPatternFromOrderNumber(orderNum).getStep(stepNum).getVolume();
@@ -16,8 +17,9 @@ SetVolumeToStepCommand::SetVolumeToStepCommand(std::weak_ptr<Module> mod, int so
 
 void SetVolumeToStepCommand::redo()
 {
+	int volume = (isFMReserved_ && vol_ < 0x80) ? (0x7f - vol_) : vol_;
 	mod_.lock()->getSong(song_).getTrack(track_).getPatternFromOrderNumber(order_)
-			.getStep(step_).setVolume(vol_);
+			.getStep(step_).setVolume(volume);
 }
 
 void SetVolumeToStepCommand::undo()